src/requirements.txt
aiofiles==0.4.0
alembic==1.4.0
aniso8601==7.0.0
astroid==2.3.3
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.20.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
databases==0.2.6
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.48.0
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.0.13
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
packaging==20.1
passlib==1.7.2
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.4
py==1.8.1
pycparser==2.19
pydantic==1.4
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.3
requests==2.22.0
Rx==1.6.1
six==1.14.0
SQLAlchemy==1.3.13
starlette==0.12.9
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.25.8
uvicorn==0.11.2
uvloop==0.14.0
wcwidth==0.1.8
websockets==8.1
wrapt==1.11.2
